// frontend/src/api/model/personDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PersonDtoEyeColor } from './personDtoEyeColor';
import type { PersonDtoHairColor } from './personDtoHairColor';
import type { LocationDto } from './locationDto';
import type { PersonDtoNationality } from './personDtoNationality';

/**
 * объект
 */
export interface PersonDto {
  id: number;
  name?: string;
  eyeColor?: PersonDtoEyeColor;
  hairColor?: PersonDtoHairColor;
  location?: LocationDto;
  birthday?: string;
  nationality?: PersonDtoNationality;
}