// frontend/src/api/model/personDtoEyeColor.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type PersonDtoEyeColor = typeof PersonDtoEyeColor[keyof typeof PersonDtoEyeColor];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PersonDtoEyeColor = {
  BLUE: 'BLUE',
  YELLOW: 'YELLOW',
  ORANGE: 'ORANGE',
} as const;