// frontend/src/api/model/personDtoHairColor.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type PersonDtoHairColor = typeof PersonDtoHairColor[keyof typeof PersonDtoHairColor];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PersonDtoHairColor = {
  BLUE: 'BLUE',
  YELLOW: 'YELLOW',
  ORANGE: 'ORANGE',
} as const;