// frontend/src/api/model/personDtoNationality.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type PersonDtoNationality = typeof PersonDtoNationality[keyof typeof PersonDtoNationality];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PersonDtoNationality = {
  USA: 'USA',
  CHINA: 'CHINA',
  VATICAN: 'VATICAN',
  THAILAND: 'THAILAND',
  NORTH_KOREA: 'NORTH_KOREA',
} as const;